Create api container base to conect with pipeline
