+++ .dbdata/ib_buffer_pool
@@ -1,22 +1,21 @@
-10,65
-10,209
-10,171
-10,239
-10,153
-10,287
-10,2
-10,3
-9,7
-9,5
-9,2
-9,3
+3,2
+3,1
+3,3
+10,173
+10,227
+10,175
+10,75
+10,309
+10,27
+10,193
+10,267
+8,5
 8,7
-8,9
-8,17
+8,10
 8,4
-8,8
-8,2
-8,3
+8,16
+38,3
+38,1
 20,3
 20,2
 20,1
@@ -31,39 +30,41 @@
 11,27
 11,26
 11,25
+11,24
 11,23
 11,22
 11,21
+11,20
 11,19
 11,17
 11,16
 11,15
 11,14
 11,13
-18,1
 11,12
-18,3
 11,11
 11,10
 11,9
 11,8
 11,7
 11,6
-14,1
 11,5
 11,4
 11,3
 11,2
 11,1
 11,0
-14,3
 10,311
 10,307
-13,1
 10,304
+18,1
+10,301
+18,3
+10,297
+14,1
+10,289
+14,3
+13,1
+10,209
 13,3
-10,276
-10,269
-10,217
-10,213
-10,201
+10,185
