+++ etl/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+
+python_files = test_*.py tests.py *_test.py
