+++ etl/requirements.txt
@@ -1,0 +1,10 @@
+pandas==1.3.5
+requests==2.27.1
+SQLAlchemy==1.4.35
+pymysql==1.0.2
+cryptography==36.0.2
+mysql-connector==2.2.9
+mysql-connector-python==8.0.29
+mysqlclient==2.1.0
+pytest==5.3.5
+datatest
