--- pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-
-python_files = test_*.py tests.py *_test.py
