--- requirements.txt
@@ -1,10 +1,0 @@
-pandas==1.3.5
-requests==2.27.1
-SQLAlchemy==1.4.35
-pymysql==1.0.2
-cryptography==36.0.2
-mysql-connector
-mysql-connector-python
-mysqlclient
-pytest
-datatest
