--- tests/test_transform.py
@@ -1,1 +1,0 @@
-import pytest
